bu hacks/requirements.txt
flask==2.0.1
tensorflow==2.8.0
Pillow==8.3.2
numpy==1.21.2 